Pull out datetime conversion to seperate file
